Use client API instead of mongo to get registrations
